requirements.txt
langchain-anthropic
langgraph
pydantic
python-dotenv
click
rich
pypdf